requirements.txt
Bootstrap_Flask==2.2.0
Flask_CKEditor==0.4.6
Flask_Login==0.6.3
Flask-Gravatar==0.5.0
Flask_WTF==1.2.1
WTForms~=3.2.1
Werkzeug~=3.1.3
Flask==2.3.2
flask_sqlalchemy==3.1.1
SQLAlchemy~=2.0.40
psycopg2-binary~=2.9.10
Flask-Login~=0.6.3
dotenv~=0.9.9
python-dotenv~=1.1.0
gunicorn==23.0.0